1.19.4 Port (#116)
* Core and base started.

Builds and runs :)

A good few things broken still and a lot of really gross hacks just to get stuff started off.

* Fixed more bugs and fixed some of datagen. Dirty for now

* Machines module builds

* Compiles with all modules enabled

* Remove some TODO marks and fix an old bug

* Resolve a range of TODOs

* Helper for tab variants

May push this into Registrate in future

* Grinding ball tooltip (needs #115's fix still)

* Missed a recipe that was generated

* Merge!

* 1.20 Dark Steel Upgrades

* Apply some optimisations

Thanks @pupnewfster for the thread!

* Apply changes from main, now I will fix merge

* Finish merge

* Apply more suggested changes

Thanks again @pupnewfster


* Disable Patchouli

* Fix!

* Re-enable JEI

* Bump forge and JEI, fixes runData

* More work, pls github actions I beg

* Port IO config widget stuff

* Re-done modrinth stuff

* Knock out a couple of TODOs

---------

+++ src/decor/java/com/enderio/decoration/client/render/entity/PaintedSandRenderer.java
@@ -33,7 +33,7 @@ public class PaintedSandRenderer extends FallingBlockRenderer {
             Level level = pEntity.getLevel();
             if (blockstate != level.getBlockState(pEntity.blockPosition()) && blockstate.getRenderShape() != RenderShape.INVISIBLE) {
                 pMatrixStack.pushPose();
-                BlockPos blockpos = new BlockPos(pEntity.getX(), pEntity.getBoundingBox().maxY, pEntity.getZ());
+                BlockPos blockpos = new BlockPos((int) pEntity.getX(), (int) pEntity.getBoundingBox().maxY, (int) pEntity.getZ());
                 pMatrixStack.translate(-0.5D, 0.0D, -0.5D);
                 BlockRenderDispatcher blockrenderdispatcher = Minecraft.getInstance().getBlockRenderer();
                 BakedModel model = blockrenderdispatcher.getBlockModel(blockstate);
